positions are uint16_t

+++ src/MovingObject.cpp
@@ -1,7 +1,7 @@
 #include <math.h>
 #include "MovingObject.h"
 
-MovingObject::MovingObject(int x_position, int y_position, double speed) : Object(x_position, y_position)
+MovingObject::MovingObject(uint16_t x_position, uint16_t y_position, double speed) : Object(x_position, y_position)
 {
     this->speed = speed;
     this->direction = 0;
@@ -10,25 +10,25 @@ MovingObject::MovingObject(int x_position, int y_position, double speed) : Objec
 void MovingObject::update(double delta)
 {
     Object::update(delta);
-    
+
     // calculate new X position
-    int x_position_offset = (int)this->speed * sin(this->direction) * delta;
-    int new_x_position = this->get_x_position() + x_position_offset;
+    uint16_t x_position_offset = (uint16_t)this->speed * sin(this->direction) * delta;
+    uint16_t new_x_position = this->get_x_position() + x_position_offset;
 
     // calculate new Y position
-    int y_position_offset = (int)this->speed * cos(this->direction) * delta;
-    int new_y_position = this->get_y_position() + y_position_offset;
+    uint16_t y_position_offset = (uint16_t)this->speed * cos(this->direction) * delta;
+    uint16_t new_y_position = this->get_y_position() + y_position_offset;
 
     // on leaving screen borders, teleport to other side
-    if(this->wrap_around_display)
+    if (this->wrap_around_display)
     {
-        if(new_x_position < 0)
+        if (new_x_position < 0)
             new_x_position = Display::WIDTH_PIXELS;
-        else if(new_x_position > Display::WIDTH_PIXELS)
+        else if (new_x_position > Display::WIDTH_PIXELS)
             new_x_position = 0;
-        if(new_y_position < 0)
+        if (new_y_position < 0)
             new_y_position = Display::HEIGHT_PIXELS;
-        else if(new_y_position > Display::HEIGHT_PIXELS)
+        else if (new_y_position > Display::HEIGHT_PIXELS)
             new_y_position = 0;
     }
 
@@ -42,13 +42,13 @@ void MovingObject::draw(Display display)
     this->undraw(display, this->previous_x_position, this->previous_y_position);
 }
 
-void MovingObject::set_x_position(int position)
+void MovingObject::set_x_position(uint16_t position)
 {
     this->previous_x_position = this->get_x_position();
     Object::set_x_position(position);
 }
 
-void MovingObject::set_y_position(int position)
+void MovingObject::set_y_position(uint16_t position)
 {
     this->previous_y_position = this->get_y_position();
     Object::set_y_position(position);
