+++ src/Player.cpp
@@ -1,6 +1,6 @@
 #include "Player.h"
 
-Player::Player(int x_position, int y_position, double speed) : MovingObject(x_position, y_position, speed)
+Player::Player(uint16_t x_position, uint16_t y_position, double speed) : MovingObject(x_position, y_position, speed)
 {
 }
 
@@ -25,7 +25,7 @@ void Player::draw(Display display)
     display.canvas.fillCircle(this->get_x_position(), this->get_y_position(), 5, ILI9341_WHITE);
 }
 
-void Player::undraw(Display display, int x_position, int y_position)
+void Player::undraw(Display display, uint16_t x_position, uint16_t y_position)
 {
     display.canvas.fillCircle(x_position, y_position, 5, display.background_colour);
 }
